% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est.auc.R
\name{est.auc}
\alias{est.auc}
\title{Sample size for estimating Area Under the ROC curve}
\usage{
est.auc(auc, alp, d)
}
\arguments{
\item{auc}{anticipated AUC of the diagnostic marker or test}

\item{alp}{level of significance or accepted level of probability of type I error}

\item{d}{Precision required on either side of the true AUC}
}
\value{
a list of total sample size based on AUC along with reporting
}
\description{
Sample size for estimating Area Under the ROC curve
}
\examples{
est.auc(auc=0.7,alp=0.05,d=0.07)
}
\references{
Hajian-Tilaki, K. (2014). Sample size estimation in diagnostic test studies of biomedical informatics. Journal of biomedical informatics, 48, 193-204.
}
