% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estRR.R
\name{estRR}
\alias{estRR}
\title{Sample size for estimation of Relative Risk}
\usage{
estRR(p0, RR, alp, prec, k)
}
\arguments{
\item{p0}{Probability of outcome among unexposed}

\item{RR}{anticipated Relative Risk (RR)}

\item{alp}{level of significance or probability of claiming
the association exists when in fact there is no association}

\item{prec}{Precision desired on either side of RR}

\item{k}{the number of unexposed for each exposed}
}
\value{
a list object, the required minimum sample size along with description for reporting
}
\description{
Relative risks are estimated in a cohort study design to assess the association of
exposure with the outcome. This function estimates the sample size needed to estimate the true relative
risk with specified precision.
}
\examples{
estRR(p0=0.2, RR=2, alp=0.05, prec=0.25, k=1)
}
\author{
R. Amala, Scientist-C, ICMR-VCRC, Puducherry  & G. Kumarapandiyan, Asst. Prof., Madras Christian College, Chennai
}
