% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testor.R
\name{testor}
\alias{testor}
\title{Sample size for testing Odds Ratio}
\usage{
testor(p0, or, alp, pwr, k)
}
\arguments{
\item{p0}{Probability of exposure among the controls}

\item{or}{Anticipated Odds Ratio}

\item{alp}{Probability of type I error}

\item{pwr}{Desired level of power}

\item{k}{ratio of number of cases to controls to cases}
}
\value{
a list object, the required minimum sample size along with description for reporting
}
\description{
When we try to associate multiple exposures to an outcome, we need
to caluclate the odds ratio (OR) of a particular exposure in the presence of other exposures
and test their relative importance in the model using a significance test based on OR. This
function computes sample size based on testing OR for a case-control study design
}
\examples{
testor(p0=0.042,or=2.5,alp=0.05,pwr=0.8,k=1)
}
\references{
Lwanga, S. K., Lemeshow, S., & World Health Organization. (1991). Sample size determination in health studies: a practical manual. World Health Organization.
}
\author{
R. Amala, Scientist-C, ICMR-VCRC, Puducherry  & G. Kumarapandiyan, Asst. Prof., Madras Christian College, Chennai
}
